Revert "change min price to double"
This reverts commit 9b22ea8d4b30152d6a1eb96e137a86c0b84ece01.

+++ constants.cs
@@ -15,9 +15,9 @@ namespace snackShack
         {
             return fileSepChar_con;
         }
-        internal static double minimumPrice()
+        internal static decimal minimumPrice()
         {
-            return (double)minPrice;
+            return (decimal)minPrice;
         }
     }
 }
